Rename votes endpoint to keep it consistent
