+++ front/app/api/idea_votes/types.ts
